Start of date handling code
